requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.3.0
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.3.0
Babel==2.15.0
beautifulsoup4==4.12.3
behave==1.2.6
bleach==6.1.0
blinker==1.8.2
bounded-pool-executor==0.0.3
cachetools==5.3.3
certifi==2024.12.14
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.7
cloudpickle==3.1.1
cloudscraper==1.2.71
cmdstanpy==1.2.5
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cramjam==2.8.3
cssselect==1.2.0
curl_cffi==0.11.3
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
dropbox==12.0.2
dune_client==1.7.8
einops==0.8.1
exceptiongroup==1.2.2
execnet==2.1.1
executing==2.0.1
fasteners==0.19
fastjsonschema==2.19.1
filelock==3.16.1
filetype==1.2.0
fonttools==4.53.0
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.10.0
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/future-0.18.2-py3-none-any.whl
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-genai==1.21.1
google-generativeai==0.8.5
googleapis-common-protos==1.63.2
greenlet==3.1.1
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.14.0
h8mail==2.5.6
holidays==0.75
httpcore==1.0.5
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipykernel==6.29.4
ipympl==0.9.6
ipython==8.18.1
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.5
jiter==0.10.0
joblib==1.5.1
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.10
kaggle==1.6.17
kiwisolver==1.4.5
langchain==0.3.25
langchain-anthropic==0.3.15
langchain-core==0.3.65
langchain-google-genai==2.1.5
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
llvmlite==0.42.0
lxml==5.2.2
lz4==4.3.3
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/macholib-1.15.2-py2.py3-none-any.whl
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
multitasking==0.0.11
mycdp==1.1.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
ndjson==0.3.1
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.2.0
notebook_shim==0.2.4
numba==0.59.1
numpy==1.26.4
opencv-python==4.10.0.84
orjson==3.10.18
ormsgpack==1.10.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.2
pandocfilters==1.5.1
parameterized==0.9.0
parse==1.20.2
parse_type==0.6.4
parso==0.8.4
pathos==0.3.2
patsy==1.0.1
pdbp==1.6.1
peewee==3.18.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.3.6
playwright==1.49.1
plotly==5.24.1
pluggy==1.5.0
ply==3.11
pox==0.3.4
ppft==1.7.6.8
pqdm==0.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
propcache==0.2.1
prophet==1.1.7
proto-plus==1.24.0
protobuf==5.27.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyee==12.0.0
Pygments==2.18.0
pynose==1.5.3
pyotp==2.9.0
pyparsing==3.1.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
PySocks==1.7.1
pytest==8.3.4
pytest-html==4.0.2
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==15.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-json-logger==2.0.7
python-slugify==8.0.4
python-snappy==0.7.2
python-telegram-bot==21.10
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
reportlab==4.2.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.18.1
rsa==4.9
sbvirtualdisplay==1.4.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
selenium==4.27.1
selenium-stealth==1.0.6
seleniumbase==4.33.12
Send2Trash==1.8.3
shap==0.48.0
six==1.17.0
slicer==0.0.8
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.41
stack-data==0.6.3
stanio==0.5.1
statsmodels==0.14.4
stone==3.3.1
streamlit==1.36.0
streamlit-custom-component==0.2
sympy==1.13.1
tabcompleter==1.4.0
tenacity==8.5.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.5.1
torchvision==0.20.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
types-Deprecated==1.2.15.20241117
types-python-dateutil==2.9.0.20240316
types-PyYAML==6.0.12.20241230
types-requests==2.32.0.20241016
types-setuptools==75.8.0.20250110
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
undetected-chromedriver==3.5.5
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.5.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
widgetsnbextension==4.0.10
wrapt==1.17.2
wsproto==1.2.0
xgboost==2.1.4
xxhash==3.5.0
yarl==1.18.3
yfinance==0.2.63
zipp==3.18.2
zstandard==0.23.0
